requirements.txt
bleach==2.1.1
bootstrap-admin==0.3.7.1
certifi==2017.7.27.1
chardet==3.0.4
coreschema==0.0.4
defusedxml==0.5.0
diff-match-patch==20121119
Django==1.11.12
django-allauth==0.33.0
django-appconf==1.0.2
django-cors-headers==2.2.0
django-crispy-forms==1.6.1
django-formtools==2.1
django-haystack==2.7.0
django-imagekit==4.0.1
django-import-export==1.0.0
django-redis==4.9.0
django-reversion==2.0.13
djangorestframework==3.8.2
emoji==0.4.5
et-xmlfile==1.0.1
future==0.15.2
gunicorn==19.7.1
html5lib==1.0b10
httplib2==0.9.2
idna==2.6
jdcal==1.4
jieba==0.39
Jinja2==2.10
Markdown==2.6.8
MarkupSafe==1.0
oauthlib==2.0.3
odfpy==1.3.6
olefile==0.44
openpyxl==2.5.3
pilkit==2.0
Pillow==4.2.1
Pygments==2.2.0
PyMySQL==0.7.11
python3-openid==3.1.0
pytz==2017.2
PyYAML==3.12
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
serializers==0.2.4
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
user-agent==0.1.9
webencodings==0.5.1
Whoosh==2.7.4
xlrd==1.1.0
xlwt==1.3.0
